优化路由类 1、匹配到规则后不再继续匹配 2、增加自定义路由参数setPatten和setPattens方法
