+++ apps/controllers/BaseController.php
@@ -1,15 +1,5 @@
 <?php
 namespace apps\controllers;
-/**
- * ----------------------
- * BaseController.class.php
- * 
- * User: jian0307@icloud.com
- * Date: 2015/7/6
- * Time: 17:00
- * ----------------------
- */
-
 class BaseController {
 
 }
